Brand Component
Created a component to display logo along with the hamburger menu. I
also created styles for each component. I'll have to make further
adjustments with the design overlay as I used the values from XD which
are usually off by a pixel or two. Lastly, I added the new components
to Header component.