Navigation And Context API
Tapped into React's Context API in order to keep track of the state of
the menu button toggle. With this implementation I was able to remove
the state logic from the components and move them into their context
file. This makes the components leaner and file organization better as
I'll be able to keep all the context logic in one place.