Backdrop Component And No Background Scrolling
Created a backdrop component and moved corresponding styles I created
in the Nav component. I originally was using portals to add the
backdrop at the root of index.html but removed it because it was not as
flexible with z-layers.

Secondly, created a close button and added it to the nav component with a
few styles.

Finally, updated the `menu-context` file and added a toggle function to
add a `noscroll` class to the body so when the menu is opened the user
will not be able to scroll the background. You see this in a lot of
websites and I personally don't like it.