Portfolio Context
Created a context to fetch portfolio data in order to provide them to a
few of my components to display Experience and Projects. Secondly, I
removed the dummy data I was using and updated with PortfolioContext.
Third, updated the keys in JobListing to match the ones I created in
DynamoDB. Finally, added the PortfolioContextProvider to App.jsx.