Skeleton Loader
Created a skeleton loader component to display while data is being
fetched. This will allow for a better user experience.

In order for the skeleton component to work properly I updated the
portfolio context to have a isLoading state so I can switch between
the loading state and data render in components.